requirements.txt
pdfplumber>=0.10.0
pandas>=2.0.0
openpyxl>=3.1.0
python-pptx>=0.6.23
PyPDF2>=3.0.0
